added code for task 1

+++ python_async_comprehension/1-async_comprehension.py
@@ -6,6 +6,6 @@ import typing
 async_generator = __import__('0-async_generator').async_generator
 
 
-async def async_comprehension() -> typing.List[int]:
+async def async_comprehension() -> typing.List[float]:
     """async list comprehension"""
     return [value async for value in async_generator()]
